+++ projects/demo/src/main/java/org/openmuc/jeebus/spine/demo/MinimalExampleDevice.java
@@ -10,7 +10,8 @@
 
 package org.openmuc.jeebus.spine.demo;
 
-import org.openmuc.jeebus.ship.api.ShipNodeConfiguration;
+import org.openmuc.jeebus.ship.api.cert.KeyStoreCertificateStorage;
+import org.openmuc.jeebus.ship.node.ShipConfig;
 import org.openmuc.jeebus.shipspine.ShipCommunication;
 import org.openmuc.jeebus.spine.api.Device;
 import org.openmuc.jeebus.spine.xsd.v1.DeviceTypeEnumType;
@@ -20,30 +21,22 @@ import static org.openmuc.jeebus.shipspine.ShipCommunication.ConnectClientsTo.TR
 
 public class MinimalExampleDevice {
     public static void main(String[] args) {
-        char[] passphrase = "yourpassphrase".toCharArray();
 
         // Configure the new SHIP node
-        ShipNodeConfiguration shipNodeConfiguration = new ShipNodeConfiguration(
-            "0.0.0.0",
-            5151,
-            "/ship/",
-            true,
-            "EXAMPLEBRAND-EEB01M3EU-001122334455",
-            "local.",
-            "Dishwasher ExampleCompany EEB01M4EU",
-            "exampleAlias",
-            "keystore.jks",
-            "CHANGEME".toCharArray(),
-            "CHANGEME".toCharArray(),
-            "CN=example name1",
-            3650
-        );
+        ShipConfig shipConfig = ShipConfig.getBuilder()
+            .withServerBindAddresses("localhost:5151")
+            .withId("JEEBUS-MINIMAL-EXAMPLE-1")
+            .withMDnsServiceInstance("jEEBus Minimal Example 1")
+            .withCertificateStorage(new KeyStoreCertificateStorage("keystore.jks"))
+            .withCertificateDistinguishedName("CN=example name1")
+            .withTrustedSkis(
+                // Here you can pre-trust remote SHIP devices per SKIs
+            )
+            .build();
 
         // Create a new SHIP communication
         ShipCommunication shipCommunication = new ShipCommunication(
-            shipNodeConfiguration
-        ).withTrustedSkis(
-            // Here you can pre-trust remote SHIP devices per SKIs
+            shipConfig
         ).withConnectClientsTo(
             TRUSTED // Configure which SHIP devices to connect to (ALL, TRUSTED, NONE)
         );
@@ -56,10 +49,7 @@ public class MinimalExampleDevice {
             // Set SHIP as the communication protocol
             .withCommunication(shipCommunication)
             // Set the SPINE device ID
-            .withId("d:_n:MinimalExample_123")
-            /* Enable the automatic SPINE DetailedDiscovery + UseCaseDiscovery of
-             * remote devices */
-            .withDiscoverDevices(true)
+            .withId("d:_n:JeebusMinimalExample_1")
             .addEntity()
                 .setType(EntityTypeEnumType.GENERIC)
                 .withUseCases(